Fix window size calculations
Now checks image width as well as height before plotting the window. Previously if the height was adjusted, no width was checking which allowed images with a certain ratio to go outside the edges of the browser window. I also increased the possible width of the window to 90% of the browser window as 70% seemed to lose a fair bit of space.